file restructure store/actions
