+++ client/store/actions/board.js
@@ -1,0 +1,1 @@
+import Board from '../../board'
