+++ client/store/actions/index.js
@@ -1,0 +1,3 @@
+export * from './users'
+export * from './board'
+export * from './actionTypes'
